+++ src/components/IngredientDetails/IngredientDetails.tsx
